Add install and update commands to Makefile
Remove "wrapped" composer and add the install and update tasks instead.

We're introducing the new, shorter "install" and "update"
tasks instead of the "install-php" and "update-php". We keep the "old"
tasks for the sake of explicitness, to be future-proof (in case we want
to install several things) and for the sake of our muscle memory.